+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased]- 
+
+## [0.11.0] - 2022-08-17
+
+### Added
+
+- **GM-135** - Add page Frequent Questions
+- **GM-132** - Contact us form
+- **GM-109** - Add component product description
+- **GM-137** - Add static institutional pages
+- **GM-131** - Add institutional menu
 
 ## [0.10.0] - 2022-08-04
 
@@ -24,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **GM-455** - Changed layout style in footer
 
 ### Added
+
 
 - **GM-397** - Filter mobile
 - **GM-398** - [PDP] Show latest questions list